Move parsing code to a separate static class
